pyproject.toml
[project]
name = "text2moo"
version = "0.1.0"
description = "Use LLM to leverage Multi Objective Optimization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastexcel>=0.14.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "openai>=1.99.8",
    "polars>=1.32.2",
    "pymoo>=0.6.1.5",
    "python-dotenv>=1.1.1",
    "xlsxwriter>=3.2.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/text2moo"]
